Rebased and updated to include SubscriptionConnector updates
