+++ test/connectors/SubscriptionConnectorSpec.scala
@@ -18,6 +18,7 @@ package connectors
 
 import assets.BaseTestConstants._
 import assets.CustomerDetailsTestConstants._
+import assets.BusinessAddressTestConstants._
 import connectors.httpParsers.ResponseHttpParser.HttpGetResult
 import mocks.MockHttp
 import models.core.SubscriptionUpdateResponseModel
@@ -67,7 +68,8 @@ class SubscriptionConnectorSpec extends TestUtil with MockHttp{
 
     "calling .updateBusinessAddress" should {
 
-      def result: Future[HttpGetResult[SubscriptionUpdateResponseModel]] = TestSubscriptionConnector.updateBusinessAddress()
+      def result: Future[HttpGetResult[SubscriptionUpdateResponseModel]] =
+        TestSubscriptionConnector.updateBusinessAddress("", addressModel)
 
       "return a SubscriptionUpdateResponseModel" in {
         await(result) shouldBe Right(SubscriptionUpdateResponseModel("12345"))
